# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Protomind API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/keywords": {
            "get": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Выводит все ключевые слова, отсортированные по времени",
                "description": "Метод для отображения списка ключевых слов",
                "operationId": "indexKeywords",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/KeywordResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Создание нового ключевого слова",
                "description": "Метод для создания нового ключевого слова",
                "operationId": "storeKeyword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KeywordStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/KeywordResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/keywords/{id}": {
            "put": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Обновление существующего ключевого слова",
                "description": "Метод для обновления существующего ключевого слова",
                "operationId": "updateKeyword",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KeywordUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeywordResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Удаление существующего ключевого слова",
                "description": "Метод для удаления ключевого слова",
                "operationId": "destroyKeyword",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ничего не найдено."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "KeywordStoreRequest": {
                "title": "Запрос на создание ключевых слов",
                "description": "Запрос, содержащий список ключевых слов для создания",
                "properties": {
                    "keywords": {
                        "description": "Список ключевых слов",
                        "type": "array",
                        "items": {
                            "properties": {
                                "title": {
                                    "description": "Заголовок ключевого слова",
                                    "type": "string"
                                },
                                "phrase": {
                                    "description": "Фраза ключевого слова",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "KeywordUpdateRequest": {
                "title": "Запрос на обновление ключевых слов",
                "description": "Запрос, содержащий список ключевых слов для обновления",
                "properties": {
                    "keywords": {
                        "description": "Список ключевых слов",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID",
                                    "type": "integer"
                                },
                                "title": {
                                    "description": "Заголовок ключевого слова",
                                    "type": "string"
                                },
                                "phrase": {
                                    "description": "Фраза ключевого слова",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "KeywordResource": {
                "title": "Keyword Resource",
                "properties": {
                    "id": {
                        "description": "ID ключевого слова",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Заголовок",
                        "type": "string"
                    },
                    "phrase": {
                        "description": "Фраза или слово",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}